Rename the validator class and remove dependency on Linq
